Entrega SERCOM

+++ common_dynamic_buffer.c
@@ -3,7 +3,7 @@
 #include <stdio.h>
 
 
-buffer_t* buffer_create(size_t length) {
+buffer_t* buffer_create(uint32_t length) {
 	buffer_t* buffer = malloc(sizeof(buffer_t));
 
 	if (buffer == NULL)
@@ -21,9 +21,9 @@ buffer_t* buffer_create(size_t length) {
 }
 
 int buffer_concatenate(buffer_t* self, const void* data_to_add, 
-					   size_t length) {
-	size_t previous_length = self->length;
-	size_t new_lenght = sizeof(char) * (length + previous_length);
+					   uint32_t length) {
+	uint32_t previous_length = self->length;
+	uint32_t new_lenght = sizeof(char) * (length + previous_length);
 	char* new_data = realloc(self->data, new_lenght);	
 	
 	if (new_data == NULL)
@@ -40,7 +40,7 @@ void* buffer_get_data(buffer_t* self){
 	return self->data;
 }
 
-size_t buffer_get_length(buffer_t* self) {
+uint32_t buffer_get_length(buffer_t* self) {
 	return self->length;
 }
